+++ src/GRA.Controllers/ProfileController.cs
@@ -1833,10 +1833,11 @@ namespace GRA.Controllers
             {
                 HasAccount = !string.IsNullOrWhiteSpace(user.Username),
                 HeadOfHouseholdId = user.HouseholdHeadUserId,
-                HouseholdCount
-                    = await _userService.FamilyMemberCountAsync(user.HouseholdHeadUserId ?? id),
+                HouseholdCount = await _userService
+                    .FamilyMemberCountAsync(user.HouseholdHeadUserId ?? id),
                 PaginateModel = paginateModel,
-                PrizeWinners = prizeList.Data
+                PrizeWinners = prizeList.Data,
+                ShowOnlyAt = await GetSiteSettingBoolAsync(SiteSettingKey.Prizes.ShowOnlyAt)
             });
         }